move onClaime to claim modal
